requirements.txt
faster-whisper
fastapi
uvicorn
loguru
python-multipart
google-generativeai
uvicorn
yt-dlp
ffmpeg-python
